produce bioregistry_ols table

produce bioregistry_ols table

Usage

bioregistry_ols_resources()

Value

data.frame

Note

This uses the `resources` method of the bioregistry module from pip to isolate resources with a non-null `ols` component.

Examples

tab = bioregistry_ols_resources()
head(tab[,1:3])
#>   contact
#> 1    <NA>
#> 2    <NA>
#> 3    <NA>
#> 4    <NA>
#> 5    <NA>
#> 6    <NA>
#>                                                                                                                                                                                                                                                                                                                                                                              description
#> 1 Alzheimer's Disease Ontology is a knowledge-based ontology that encompasses varieties of concepts related to Alzheimer'S Disease, foundamentally structured by upper level Basic Formal Ontology(BFO). This Ontology is enriched by the interrelational entities that demonstrate the nextwork of the understanding on Alzheimer's disease and can be readily applied for text mining.
#> 2                                                                                                                                                                                                                                                                                   AEO is an ontology of anatomical structures that expands CARO, the Common Anatomy Reference Ontology
#> 3                                                                                                                                                                                                        The Adverse Event Reporting Ontology (AERO) is an ontology aimed at supporting clinicians at the time of data entry, increasing quality and accuracy of reported adverse events
#> 4                                                                                                                                                                                                                                                                                                                                                        Allotrope Merged Ontology Suite
#> 5                                                                                                                                                                                                                                                                                    AgrO is an ontlogy for representing agronomic practices, techniques, variables and related entities
#> 6                                                                                                                                                  The ontology for the Anatomy of the Insect SkeletoMuscular system (AISM) contains terms used to describe the cuticle - as a single anatomical structure - and the skeletal muscle system, to be used in insect biodiversity research.
#>                              download_owl
#> 1  http://purl.obolibrary.org/obo/ado.owl
#> 2  http://purl.obolibrary.org/obo/aeo.owl
#> 3                                    <NA>
#> 4                                    <NA>
#> 5 http://purl.obolibrary.org/obo/agro.owl
#> 6 http://purl.obolibrary.org/obo/aism.owl