obtain list of names of a set of subclasses/children

obtain list of names of a set of subclasses/children

Usage

children_names(sclist)

Arguments

sclist

output of `subclasses`

Value

list of vectors of character()

Note

non-entities are removed and names are extracted

Examples

pa = get_ordo_owl_path()
o2 = try(reticulate::import("owlready2"), silent=TRUE)
if (!inherits(o2, "try-error")) {
 orde = setup_entities(pa)
 al = subclasses(orde[100:120])
 children_names(al)
}
#> $Orphanet_98131
#> list()
#> 
#> $Orphanet_100072
#> list()
#> 
#> $Orphanet_C052
#>   [1] "Orphanet_100072" "Orphanet_100990" "Orphanet_101036" "Orphanet_101052"
#>   [5] "Orphanet_101356" "Orphanet_102069" "Orphanet_1032"   "Orphanet_103912"
#>   [9] "Orphanet_103915" "Orphanet_103916" "Orphanet_103917" "Orphanet_1044"  
#>  [13] "Orphanet_1057"   "Orphanet_1088"   "Orphanet_1115"   "Orphanet_1137"  
#>  [17] "Orphanet_1139"   "Orphanet_1153"   "Orphanet_1155"   "Orphanet_1167"  
#>  [21] "Orphanet_1211"   "Orphanet_1235"   "Orphanet_1239"   "Orphanet_1249"  
#>  [25] "Orphanet_1250"   "Orphanet_1256"   "Orphanet_1258"   "Orphanet_1260"  
#>  [29] "Orphanet_1339"   "Orphanet_1376"   "Orphanet_137871" "Orphanet_137923"
#>  [33] "Orphanet_1396"   "Orphanet_1417"   "Orphanet_1420"   "Orphanet_1421"  
#>  [37] "Orphanet_1428"   "Orphanet_1434"   "Orphanet_1492"   "Orphanet_1499"  
#>  [41] "Orphanet_1526"   "Orphanet_1530"   "Orphanet_1533"   "Orphanet_1534"  
#>  [45] "Orphanet_1575"   "Orphanet_1577"   "Orphanet_158796" "Orphanet_158799"
#>  [49] "Orphanet_1611"   "Orphanet_1625"   "Orphanet_162521" "Orphanet_163678"
#>  [53] "Orphanet_163898" "Orphanet_163908" "Orphanet_163914" "Orphanet_163924"
#>  [57] "Orphanet_163988" "Orphanet_1654"   "Orphanet_1664"   "Orphanet_1680"  
#>  [61] "Orphanet_171201" "Orphanet_171208" "Orphanet_171215" "Orphanet_171860"
#>  [65] "Orphanet_1739"   "Orphanet_178330" "Orphanet_1789"   "Orphanet_1800"  
#>  [69] "Orphanet_1819"   "Orphanet_1823"   "Orphanet_1844"   "Orphanet_1849"  
#>  [73] "Orphanet_1864"   "Orphanet_1939"   "Orphanet_1956"   "Orphanet_2005"  
#>  [77] "Orphanet_2033"   "Orphanet_2042"   "Orphanet_2054"   "Orphanet_206616"
#>  [81] "Orphanet_206619" "Orphanet_208600" "Orphanet_209886" "Orphanet_2099"  
#>  [85] "Orphanet_210592" "Orphanet_2112"   "Orphanet_2120"   "Orphanet_2129"  
#>  [89] "Orphanet_213741" "Orphanet_2156"   "Orphanet_2161"   "Orphanet_217052"
#>  [93] "Orphanet_217601" "Orphanet_2190"   "Orphanet_221106" "Orphanet_221150"
#>  [97] "Orphanet_2245"   "Orphanet_2258"   "Orphanet_225968" "Orphanet_2267"  
#> [101] "Orphanet_2284"   "Orphanet_228418" "Orphanet_2343"   "Orphanet_238691"
#> [105] "Orphanet_240843" "Orphanet_240847" "Orphanet_240849" "Orphanet_240851"
#> [109] "Orphanet_240853" "Orphanet_240855" "Orphanet_240857" "Orphanet_240859"
#> [113] "Orphanet_240861" "Orphanet_240865" "Orphanet_240873" "Orphanet_240875"
#> [117] "Orphanet_240877" "Orphanet_240879" "Orphanet_240881" "Orphanet_240883"
#> [121] "Orphanet_240889" "Orphanet_240891" "Orphanet_240893" "Orphanet_240895"
#> [125] "Orphanet_240897" "Orphanet_240899" "Orphanet_240901" "Orphanet_240903"
#> [129] "Orphanet_240907" "Orphanet_240909" "Orphanet_240911" "Orphanet_240913"
#> [133] "Orphanet_240915" "Orphanet_240917" "Orphanet_240919" "Orphanet_240923"
#> [137] "Orphanet_240925" "Orphanet_240927" "Orphanet_240929" "Orphanet_240931"
#> [141] "Orphanet_240933" "Orphanet_240937" "Orphanet_240939" "Orphanet_240941"
#> [145] "Orphanet_240943" "Orphanet_240945" "Orphanet_240949" "Orphanet_240951"
#> [149] "Orphanet_240953" "Orphanet_240955" "Orphanet_240957" "Orphanet_240959"
#> [153] "Orphanet_240961" "Orphanet_240963" "Orphanet_240965" "Orphanet_240967"
#> [157] "Orphanet_240969" "Orphanet_240971" "Orphanet_240973" "Orphanet_240975"
#> [161] "Orphanet_240977" "Orphanet_240979" "Orphanet_240981" "Orphanet_240983"
#> [165] "Orphanet_240985" "Orphanet_240987" "Orphanet_240989" "Orphanet_240991"
#> [169] "Orphanet_240993" "Orphanet_240995" "Orphanet_240997" "Orphanet_240999"
#> [173] "Orphanet_241001" "Orphanet_241003" "Orphanet_241005" "Orphanet_241007"
#> [177] "Orphanet_241009" "Orphanet_241011" "Orphanet_241013" "Orphanet_241015"
#> [181] "Orphanet_241017" "Orphanet_241019" "Orphanet_241021" "Orphanet_241023"
#> [185] "Orphanet_241025" "Orphanet_241027" "Orphanet_241029" "Orphanet_241031"
#> [189] "Orphanet_241033" "Orphanet_241035" "Orphanet_241037" "Orphanet_241039"
#> [193] "Orphanet_241041" "Orphanet_241045" "Orphanet_2454"   "Orphanet_2458"  
#> [197] "Orphanet_2474"   "Orphanet_247871" "Orphanet_248305" "Orphanet_2507"  
#> [201] "Orphanet_251316" "Orphanet_251633" "Orphanet_251891" "Orphanet_2535"  
#> [205] "Orphanet_2543"   "Orphanet_255117" "Orphanet_255199" "Orphanet_255225"
#> [209] "Orphanet_2580"   "Orphanet_2601"   "Orphanet_261512" "Orphanet_2621"  
#> [213] "Orphanet_2626"   "Orphanet_2641"   "Orphanet_2650"   "Orphanet_2675"  
#> [217] "Orphanet_2677"   "Orphanet_2679"   "Orphanet_268871" "Orphanet_269200"
#> [221] "Orphanet_2694"   "Orphanet_2705"   "Orphanet_2706"   "Orphanet_2708"  
#> [225] "Orphanet_2716"   "Orphanet_2742"   "Orphanet_275534" "Orphanet_276402"
#> [229] "Orphanet_2765"   "Orphanet_278"    "Orphanet_280569" "Orphanet_2823"  
#> [233] "Orphanet_284113" "Orphanet_28455"  "Orphanet_2860"   "Orphanet_2861"  
#> [237] "Orphanet_2864"   "Orphanet_2894"   "Orphanet_289527" "Orphanet_2925"  
#> [241] "Orphanet_293190" "Orphanet_294990" "Orphanet_294992" "Orphanet_294994"
#> [245] "Orphanet_294996" "Orphanet_294998" "Orphanet_295006" "Orphanet_295008"
#> [249] "Orphanet_295010" "Orphanet_300337" "Orphanet_3043"   "Orphanet_3046"  
#> [253] "Orphanet_3050"   "Orphanet_3061"   "Orphanet_3062"   "Orphanet_3064"  
#> [257] "Orphanet_306539" "Orphanet_306566" "Orphanet_306574" "Orphanet_3067"  
#> [261] "Orphanet_3122"   "Orphanet_3128"   "Orphanet_3133"   "Orphanet_314946"
#> [265] "Orphanet_3160"   "Orphanet_319314" "Orphanet_3215"   "Orphanet_3228"  
#> [269] "Orphanet_3229"   "Orphanet_3267"   "Orphanet_3284"   "Orphanet_329206"
#> [273] "Orphanet_329252" "Orphanet_329341" "Orphanet_329998" "Orphanet_330009"
#> [277] "Orphanet_3311"   "Orphanet_3313"   "Orphanet_3315"   "Orphanet_3331"  
#> [281] "Orphanet_3340"   "Orphanet_3354"   "Orphanet_3357"   "Orphanet_3360"  
#> [285] "Orphanet_3362"   "Orphanet_3419"   "Orphanet_34527"  "Orphanet_35061" 
#> [289] "Orphanet_35064"  "Orphanet_35065"  "Orphanet_35098"  "Orphanet_35123" 
#> [293] "Orphanet_352482" "Orphanet_352497" "Orphanet_352504" "Orphanet_352704"
#> [297] "Orphanet_35688"  "Orphanet_363629" "Orphanet_369894" "Orphanet_369902"
#> [301] "Orphanet_391658" "Orphanet_412220" "Orphanet_413690" "Orphanet_415675"
#> [305] "Orphanet_423781" "Orphanet_430"    "Orphanet_435"    "Orphanet_435623"
#> [309] "Orphanet_440"    "Orphanet_443301" "Orphanet_443304" "Orphanet_443307"
#> [313] "Orphanet_443310" "Orphanet_443313" "Orphanet_443316" "Orphanet_443319"
#> [317] "Orphanet_443322" "Orphanet_443325" "Orphanet_443328" "Orphanet_447792"
#> [321] "Orphanet_459353" "Orphanet_46489"  "Orphanet_466732" "Orphanet_477668"
#> [325] "Orphanet_477697" "Orphanet_480773" "Orphanet_498700" "Orphanet_50920" 
#> [329] "Orphanet_51013"  "Orphanet_519394" "Orphanet_521268" "Orphanet_57194" 
#> [333] "Orphanet_58220"  "Orphanet_63440"  "Orphanet_669"    "Orphanet_67037" 
#> [337] "Orphanet_70470"  "Orphanet_71269"  "Orphanet_71270"  "Orphanet_719"   
#> [341] "Orphanet_73274"  "Orphanet_75501"  "Orphanet_77241"  "Orphanet_77242" 
#> [345] "Orphanet_77304"  "Orphanet_775"    "Orphanet_788"    "Orphanet_79132" 
#> [349] "Orphanet_79344"  "Orphanet_83648"  "Orphanet_84096"  "Orphanet_85281" 
#> [353] "Orphanet_85318"  "Orphanet_85331"  "Orphanet_86917"  "Orphanet_89839" 
#> [357] "Orphanet_89840"  "Orphanet_89841"  "Orphanet_89845"  "Orphanet_90022" 
#> [361] "Orphanet_90055"  "Orphanet_90070"  "Orphanet_90079"  "Orphanet_90287" 
#> [365] "Orphanet_90339"  "Orphanet_90345"  "Orphanet_91128"  "Orphanet_91133" 
#> [369] "Orphanet_91353"  "Orphanet_91365"  "Orphanet_91415"  "Orphanet_93313" 
#> [373] "Orphanet_93594"  "Orphanet_93937"  "Orphanet_93942"  "Orphanet_93951" 
#> [377] "Orphanet_93955"  "Orphanet_93956"  "Orphanet_93957"  "Orphanet_93961" 
#> [381] "Orphanet_93962"  "Orphanet_93963"  "Orphanet_93975"  "Orphanet_94061" 
#> [385] "Orphanet_94095"  "Orphanet_953"    "Orphanet_95426"  "Orphanet_95449" 
#> [389] "Orphanet_95458"  "Orphanet_95484"  "Orphanet_95614"  "Orphanet_95621" 
#> [393] "Orphanet_95622"  "Orphanet_95623"  "Orphanet_95625"  "Orphanet_95701" 
#> [397] "Orphanet_95721"  "Orphanet_96136"  "Orphanet_96152"  "Orphanet_96164" 
#> [401] "Orphanet_97342"  "Orphanet_97599"  "Orphanet_97927"  "Orphanet_98345" 
#> [405] "Orphanet_98593"  "Orphanet_98866"  "Orphanet_98941"  "Orphanet_99022" 
#> [409] "Orphanet_99085"  "Orphanet_99086"  "Orphanet_99088"  "Orphanet_99096" 
#> [413] "Orphanet_99097"  "Orphanet_99126"  "Orphanet_99134"  "Orphanet_99143" 
#> [417] "Orphanet_99146"  "Orphanet_99648"  "Orphanet_99649"  "Orphanet_99654" 
#> [421] "Orphanet_99662"  "Orphanet_99663"  "Orphanet_99664"  "Orphanet_99666" 
#> [425] "Orphanet_99706"  "Orphanet_99877"  "Orphanet_99878"  "Orphanet_99972" 
#> [429] "Orphanet_99973" 
#> 
#> $Orphanet_97330
#> list()
#> 
#> $Orphanet_100073
#> list()
#> 
#> $Orphanet_182086
#>  [1] "Orphanet_209007" "Orphanet_207038" "Orphanet_206613" "Orphanet_209016"
#>  [5] "Orphanet_208974" "Orphanet_209010" "Orphanet_208984" "Orphanet_209013"
#>  [9] "Orphanet_209019" "Orphanet_221109" "Orphanet_221114"
#> 
#> $Orphanet_100075
#> list()
#> 
#> $Orphanet_481508
#> [1] "Orphanet_423975"
#> 
#> $Orphanet_63443
#> [1] "Orphanet_423771"
#> 
#> $Orphanet_100076
#> list()
#> 
#> $Orphanet_423975
#> [1] "Orphanet_100076" "Orphanet_100077"
#> 
#> $Orphanet_100077
#> list()
#> 
#> $Orphanet_100078
#> list()
#> 
#> $Orphanet_100079
#> list()
#> 
#> $Orphanet_423982
#> list()
#> 
#> $Orphanet_100080
#> list()
#> 
#> $Orphanet_423991
#> list()
#> 
#> $Orphanet_100081
#> list()
#> 
#> $Orphanet_423998
#> list()
#> 
#> $Orphanet_100082
#> list()
#> 
#> $Orphanet_424010
#> [1] "Orphanet_424013"
#>