as in Bakken et al. (2017 PMID 29322913) create gene signatures for k cell types, each of which fails to express all but one gene in a set of k genes

as in Bakken et al. (2017 PMID 29322913) create gene signatures for k cell types, each of which fails to express all but one gene in a set of k genes

cyclicSigset(
  idvec,
  conds = c("hasExp", "lacksExp"),
  tags = paste0("CL:X", 1:length(idvec))
)

Arguments

idvec
character vector of identifiers, must have names() set to identify cells bearing genes

conds
character(2) tokens used to indicate condition to which signature element contributes

tags
character vector of cell-type identifiers; for Cell Ontology use CL: as prefix, one element for each element of idvec

Value

a long data.frame

Examples

sigels = c("CL:X01"="GRIK3", "CL:X02"="NTNG1", "CL:X03"="BAGE2", 
        "CL:X04"="MC4R", "CL:X05"="PAX6", "CL:X06"="TSPAN12", "CL:X07"="hSHISA8", 
     "CL:X08"="SNCG", "CL:X09"="ARHGEF28", "CL:X10"="EGF")
sigdf = cyclicSigset(sigels)
head(sigdf)

#>       gene   type     cond
#> 1 ARHGEF28 CL:X09   hasExp
#> 2    GRIK3 CL:X09 lacksExp
#> 3    NTNG1 CL:X09 lacksExp
#> 4    BAGE2 CL:X09 lacksExp
#> 5     MC4R CL:X09 lacksExp
#> 6     PAX6 CL:X09 lacksExp