extract class relationship graph from JSON representation of OWL

extract class relationship graph from JSON representation of OWL

Usage

jowl2classgraph(
  jsonpath,
  dropstrings = c("http://www.ifomis.org/bfo/1.1/snap#",
    "http://purl.org/autism-ontology/1.0/autism-rules.owl#")
)

Arguments

jsonpath

character(1) path to JSON, typically generated by java robot applied to owl

dropstrings

character(), strings to be excised from class names

Value

graphNEL with edgemode 'directed'

Examples

if (!requireNamespace("graph")) stop("install graph package from Bioconductor to use this function")
jpath = system.file("json", "aut.json.gz", package="ontoProc")
cg = jowl2classgraph(jpath,
  dropstrings = "http://purl.org/autism-ontology/1.0/autism-rules.owl#")
head(graph::nodes(cg))
#> [1] "Intense Interests"                                              
#> [2] "Understanding OthersUnderstanding Expectations"                 
#> [3] "Only occasional complex mannerisms or stereotyped body movement"
#> [4] "Autism High Level Visualizer"                                   
#> [5] "Emotional Regulation and Control"                               
#> [6] "Pathologic FunctionPathologic NOS"