Produce a data.frame with a set of naive terms mapped to all matching ontology ids and their formal terms

Produce a data.frame with a set of naive terms mapped to all matching ontology ids and their formal terms

liberalMap(terms, onto, useAgrep = FALSE, ...)

Arguments

terms
character() vector, can use grep-compatible regular expressions

onto
an instance of ontologyIndex::ontology_index

useAgrep
logical(1) if TRUE, agrep will be used

...
passed to agrep if used

Value

a data.frame

Examples

cands = c("astrocyte$", "oligodendrocyte", "oligodendrocyte precursor",
   "neoplastic", "^neuron$", "^vascular", "badterm")
co = ontoProc::getCellOnto()
liberalMap(cands, co)

#>              input         ontoid
#> 1       astrocyte$     CL:0000127
#> 2       astrocyte$     CL:0002626
#> 3       astrocyte$     CL:0002627
#> 4  oligodendrocyte     CL:0000128
#> 5  oligodendrocyte     CL:0002453
#> 6  oligodendrocyte     CL:2000025
#> 8       neoplastic     CL:0001063
#> 9       neoplastic   PATO:0002011
#> 10      neoplastic   PATO:0002097
#> 11        ^neuron$     CL:0000540
#> 12       ^vascular     CL:0000359
#> 13       ^vascular     CL:0005022
#> 14       ^vascular   PR:000001971
#> 15       ^vascular   PR:000002112
#> 16       ^vascular   PR:000048096
#> 17       ^vascular      PR:P35918
#> 18       ^vascular      PR:P35968
#> 19       ^vascular UBERON:0005629
#> 20       ^vascular UBERON:0006965
#> 21       ^vascular UBERON:0007798
#> 22       ^vascular UBERON:0009536
#> 23       ^vascular UBERON:0009537
#> 24         badterm        badterm
#>                                                                                  term
#> 1                                                                           astrocyte
#> 2                                                                  immature astrocyte
#> 3                                                                    mature astrocyte
#> 4                                                                     oligodendrocyte
#> 5                                                      oligodendrocyte precursor cell
#> 6                                                         spinal cord oligodendrocyte
#> 8                                                                     neoplastic cell
#> 9                                                                          neoplastic
#> 10                                                              neoplastic, malignant
#> 11                                                                             neuron
#> 12                                             vascular associated smooth muscle cell
#> 13                                                           vascular lymphangioblast
#> 14                                        vascular endothelial growth factor receptor
#> 15                                      vascular endothelial growth factor receptor 2
#> 16 vascular endothelial growth factor receptor 2, signal peptide removed form (human)
#> 17                              vascular endothelial growth factor receptor 2 (mouse)
#> 18                              vascular endothelial growth factor receptor 2 (human)
#> 19                                                                    vascular plexus
#> 20                                                                      vascular cord
#> 21                                                                    vascular system
#> 22                                                      vascular element of left lung
#> 23                                                     vascular element of right lung
#> 24                                                                               <NA>