packDesc2023: overview of ontoProc resources

packDesc2023: overview of ontoProc resources

data(packDesc2023)

Format

data.frame instance

Note

Brief survey of functions available to load serialized ontology_index instances imported from OBO. Focus is on versions added in 2023. Several manual interventions were needed -- cellosaurus was too large to use the script in inst/scripts/desc.R, and a number of ontologies do not have 2023 versions.

Examples

data(packDesc2023)
head(packDesc2023)
#>   X         name nclass nprop nroots                                  datav
#> 1 1         caro    751    57     30 caro/releases/2022-02-18/caro-full.owl
#> 2 2 cellLineOnto  43596   204    110                                   <NA>
#> 3 3     cellOnto  17371   229    102                    releases/2023-02-15
#> 4 4  cellosaurus 144570    NA 144571                                   44.0
#> 5 5   chebi_full 163735    34     13                                    218
#> 6 6   chebi_lite 163735    32     13                                    218
#>   fmtv
#> 1  1.2
#> 2  1.2
#> 3  1.2
#> 4  1.2
#> 5  1.2
#> 6  1.2