Probe the cache for entries found by bfcquery using the given query.

Probe the cache for entries found by bfcquery using the given query.

Usage

quickOnto(query, cache = BiocFileCache::BiocFileCache(), qans.only = FALSE)

Arguments

query

character(1), will be used by BiocFileCache::bfcquery

cache

instance of class(BiocFileCache::BiocFileCache())

qans.only

logical, defaulting to FALSE, if TRUE, just return the bfcquery result.

Note

Take only the last entry found if there are multiple hits. Use `setup_entities2` to transform owl to ontologyIndex and cache the result if this is absent for the owl identified by query.

Examples

aeo = quickOnto("aeo.owl")
#> multiple hits, returning last
str(aeo)
#> List of 6
#>  $ id       : Named chr [1:250] "AEO_0000013" "CARO_0000043" "AEO_0000078" "AEO_0000221" ...
#>   ..- attr(*, "names")= chr [1:250] "AEO_0000013" "CARO_0000043" "AEO_0000078" "AEO_0000221" ...
#>  $ name     : Named chr [1:250] "single-cell tissue" "simple tissue" "lumen of tube" "open anatomical space" ...
#>   ..- attr(*, "names")= chr [1:250] "AEO_0000013" "CARO_0000043" "AEO_0000078" "AEO_0000221" ...
#>  $ parents  :List of 250
#>   ..$ AEO_0000013 : chr "CARO_0000043"
#>   ..$ CARO_0000043: chr "CARO_0000003"
#>   ..$ AEO_0000078 : chr "AEO_0000221"
#>   ..$ AEO_0000221 : chr "CARO_0000005"
#>   ..$ AEO_0000079 : chr "AEO_0000222"
#>   ..$ AEO_0000222 : chr "CARO_0000005"
#>   ..$ AEO_0000080 : chr "CARO_0000010"
#>   ..$ CARO_0000010: chr "CARO_0000007"
#>   ..$ AEO_0000081 : chr "CARO_0000043"
#>   ..$ AEO_0000082 : chr "CARO_0000055"
#>   ..$ CARO_0000055: chr "CARO_0000003"
#>   ..$ AEO_0000083 : chr "AEO_0000082"
#>   ..$ CL_0000138  : chr "CL_0000667"
#>   ..$ CL_0000058  : chr [1:2] "CL_0000055" "CL_0000667"
#>   ..$ CL_0000062  : chr [1:2] "CL_0000055" "CL_0002320"
#>   ..$ AEO_0000168 : chr "CARO_0000011"
#>   ..$ AEO_0000085 : chr "AEO_0000082"
#>   ..$ AEO_0000086 : chr "AEO_0000083"
#>   ..$ AEO_0000087 : chr "AEO_0000081"
#>   ..$ AEO_0000090 : chr "AEO_0000081"
#>   ..$ CL_0000327  : chr [1:2] "CL_0000151" "CL_0000499"
#>   ..$ CL_0000057  : chr "CL_0002320"
#>   ..$ AEO_0000091 : chr "AEO_0000081"
#>   ..$ AEO_0000092 : chr "AEO_0000081"
#>   ..$ CL_0002320  : chr "CL_0002371"
#>   ..$ AEO_0000093 : chr "CARO_0000041"
#>   ..$ CARO_0000041: chr "CARO_0000054"
#>   ..$ AEO_0000094 : chr "CARO_0000024"
#>   ..$ CARO_0000024: chr "CARO_0000003"
#>   ..$ AEO_0000095 : chr "CARO_0000055"
#>   ..$ AEO_0000096 : chr "CARO_0000055"
#>   ..$ AEO_0000097 : chr "AEO_0000096"
#>   ..$ AEO_0000098 : chr "AEO_0000096"
#>   ..$ AEO_0000099 : chr "CARO_0000040"
#>   ..$ CARO_0000040: chr "CARO_0000003"
#>   ..$ AEO_0000100 : chr "CARO_0000040"
#>   ..$ AEO_0000103 : chr "CARO_0000032"
#>   ..$ CARO_0000032: chr "CARO_0000003"
#>   ..$ AEO_0000106 : chr "CARO_0000032"
#>   ..$ AEO_0000108 : chr "CARO_0000032"
#>   ..$ AEO_0000109 : chr "AEO_0000193"
#>   ..$ AEO_0000193 : chr "CARO_0000032"
#>   ..$ AEO_0000111 : chr "CARO_0000070"
#>   ..$ CARO_0000070: chr "CARO_0000073"
#>   ..$ AEO_0000114 : chr "CARO_0000066"
#>   ..$ CARO_0000066: chr "CARO_0000043"
#>   ..$ CL_0000068  : chr [1:2] "CL_0000066" "CL_0002371"
#>   ..$ AEO_0000115 : chr "CARO_0000066"
#>   ..$ CL_0000075  : chr [1:2] "CL_0000066" "CL_0002371"
#>   ..$ AEO_0000117 : chr "AEO_0000115"
#>   ..$ AEO_0000118 : chr "AEO_0000114"
#>   ..$ AEO_0000119f: chr "CARO_0000066"
#>   ..$ AEO_0000122 : chr "CARO_0000043"
#>   ..$ AEO_0000123 : chr "CARO_0000043"
#>   ..$ AEO_0000125 : chr "CARO_0000003"
#>   ..$ CARO_0000003: chr "CARO_0000006"
#>   ..$ AEO_0000126 : chr "CARO_0000012"
#>   ..$ CARO_0000012: chr "CARO_0000003"
#>   ..$ AEO_0000127 : chr "AEO_0000125"
#>   ..$ CL_0000445  : chr "CL_0000144"
#>   ..$ AEO_0000129 : chr "AEO_0000125"
#>   ..$ AEO_0000131 : chr "AEO_0000125"
#>   ..$ CL_0000034  : chr "CL_0000144"
#>   ..$ AEO_0000132 : chr "AEO_0000125"
#>   ..$ AEO_0000135 : chr "AEO_0000123"
#>   ..$ AEO_0000136 : chr "AEO_0001012"
#>   ..$ AEO_0001012 : chr "AEO_0000123"
#>   ..$ CL_0000127  : chr "CL_0000128"
#>   ..$ CL_0000129  : chr "CL_0000243"
#>   ..$ CL_0000128  : chr "CL_0000126"
#>   ..$ AEO_0000137 : chr "AEO_0000123"
#>   ..$ AEO_0000138 : chr "AEO_0000123"
#>   ..$ CL_0000526  : chr "CL_0000540"
#>   ..$ CL_0000527  : chr "CL_0000540"
#>   ..$ AEO_0000139 : chr "AEO_0001011"
#>   ..$ AEO_0001011 : chr "AEO_0000123"
#>   ..$ AEO_0000140 : chr "AEO_0000122"
#>   ..$ CL_0000737  : chr "CL_0000187"
#>   ..$ AEO_0000141 : chr "AEO_0000122"
#>   ..$ CL_0000192  : chr "CL_0000187"
#>   ..$ AEO_0000142 : chr "AEO_0000122"
#>   ..$ CL_0000746  : chr "CL_0000737"
#>   ..$ AEO_0000143 : chr "AEO_0000141"
#>   ..$ AEO_0000144 : chr "AEO_0000140"
#>   ..$ AEO_0000145 : chr "CARO_0000043"
#>   ..$ AEO_0000146 : chr "AEO_0000145"
#>   ..$ AEO_0000147 : chr "AEO_0000148"
#>   ..$ AEO_0000148 : chr [1:2] "AEO_0000146" "AEO_0001016"
#>   ..$ AEO_0001016 : chr "AEO_0000125"
#>   ..$ CL_0000335  : chr "CL_0000134"
#>   ..$ AEO_0000149 : chr "AEO_0000148"
#>   ..$ CL_0000680  : chr [1:2] "CL_0000055" "CL_0002371"
#>   ..$ AEO_0000150 : chr "AEO_0000148"
#>   ..$ AEO_0000151 : chr "AEO_0000145"
#>   ..$ AEO_0000152 : chr "AEO_0000145"
#>   ..$ CL_0000134  : chr "CL_0000063"
#>   ..$ AEO_0000153 : chr "AEO_0000187"
#>   ..$ AEO_0000187 : chr "CARO_0000055"
#>   ..$ AEO_0000154 : chr "CARO_0000032"
#>   .. [list output truncated]
#>  $ children :List of 250
#>   ..$ AEO_0000013 : chr(0) 
#>   ..$ CARO_0000043: chr [1:10] "AEO_0000013" "AEO_0000081" "CARO_0000066" "AEO_0000122" ...
#>   ..$ AEO_0000078 : chr(0) 
#>   ..$ AEO_0000221 : chr [1:2] "AEO_0000078" "AEO_0000215"
#>   ..$ AEO_0000079 : chr(0) 
#>   ..$ AEO_0000222 : chr "AEO_0000079"
#>   ..$ AEO_0000080 : chr(0) 
#>   ..$ CARO_0000010: chr [1:5] "AEO_0000080" "AEO_0000161" "AEO_0000162" "AEO_0000204" ...
#>   ..$ AEO_0000081 : chr [1:6] "AEO_0000087" "AEO_0000090" "AEO_0000091" "AEO_0000092" ...
#>   ..$ AEO_0000082 : chr [1:2] "AEO_0000083" "AEO_0000085"
#>   ..$ CARO_0000055: chr [1:10] "AEO_0000082" "AEO_0000095" "AEO_0000096" "AEO_0000187" ...
#>   ..$ AEO_0000083 : chr "AEO_0000086"
#>   ..$ CL_0000138  : chr(0) 
#>   ..$ CL_0000058  : chr(0) 
#>   ..$ CL_0000062  : chr(0) 
#>   ..$ AEO_0000168 : chr(0) 
#>   ..$ AEO_0000085 : chr(0) 
#>   ..$ AEO_0000086 : chr(0) 
#>   ..$ AEO_0000087 : chr [1:2] "AEO_0000179" "AEO_0000180"
#>   ..$ AEO_0000090 : chr(0) 
#>   ..$ CL_0000327  : chr [1:2] "CL_0000658" "CL_0000667"
#>   ..$ CL_0000057  : chr(0) 
#>   ..$ AEO_0000091 : chr(0) 
#>   ..$ AEO_0000092 : chr(0) 
#>   ..$ CL_0002320  : chr [1:4] "CL_0000062" "CL_0000057" "CL_0000136" "CL_0000499"
#>   ..$ AEO_0000093 : chr(0) 
#>   ..$ CARO_0000041: chr [1:2] "AEO_0000093" "AEO_0000211"
#>   ..$ AEO_0000094 : chr "AEO_0000173"
#>   ..$ CARO_0000024: chr [1:3] "AEO_0000094" "CARO_0000072" "CARO_0000074"
#>   ..$ AEO_0000095 : chr(0) 
#>   ..$ AEO_0000096 : chr [1:2] "AEO_0000097" "AEO_0000098"
#>   ..$ AEO_0000097 : chr(0) 
#>   ..$ AEO_0000098 : chr(0) 
#>   ..$ AEO_0000099 : chr(0) 
#>   ..$ CARO_0000040: chr [1:5] "AEO_0000099" "AEO_0000100" "AEO_0000188" "AEO_0001003" ...
#>   ..$ AEO_0000100 : chr(0) 
#>   ..$ AEO_0000103 : chr(0) 
#>   ..$ CARO_0000032: chr [1:5] "AEO_0000103" "AEO_0000106" "AEO_0000108" "AEO_0000193" ...
#>   ..$ AEO_0000106 : chr(0) 
#>   ..$ AEO_0000108 : chr(0) 
#>   ..$ AEO_0000109 : chr(0) 
#>   ..$ AEO_0000193 : chr [1:3] "AEO_0000109" "AEO_0000172" "AEO_0001004"
#>   ..$ AEO_0000111 : chr(0) 
#>   ..$ CARO_0000070: chr "AEO_0000111"
#>   ..$ AEO_0000114 : chr [1:2] "AEO_0000118" "AEO_0000207"
#>   ..$ CARO_0000066: chr [1:9] "AEO_0000114" "AEO_0000115" "AEO_0000119f" "CARO_0000069" ...
#>   ..$ CL_0000068  : chr(0) 
#>   ..$ AEO_0000115 : chr "AEO_0000117"
#>   ..$ CL_0000075  : chr(0) 
#>   ..$ AEO_0000117 : chr(0) 
#>   ..$ AEO_0000118 : chr [1:2] "AEO_0000219" "AEO_0001002"
#>   ..$ AEO_0000119f: chr "AEO_0001015"
#>   ..$ AEO_0000122 : chr [1:3] "AEO_0000140" "AEO_0000141" "AEO_0000142"
#>   ..$ AEO_0000123 : chr [1:5] "AEO_0000135" "AEO_0001012" "AEO_0000137" "AEO_0000138" ...
#>   ..$ AEO_0000125 : chr [1:13] "AEO_0000127" "AEO_0000129" "AEO_0000131" "AEO_0000132" ...
#>   ..$ CARO_0000003: chr [1:13] "CARO_0000043" "CARO_0000055" "CARO_0000024" "CARO_0000040" ...
#>   ..$ AEO_0000126 : chr(0) 
#>   ..$ CARO_0000012: chr [1:6] "AEO_0000126" "AEO_0000169" "AEO_0000194" "CARO_0000048" ...
#>   ..$ AEO_0000127 : chr(0) 
#>   ..$ CL_0000445  : chr(0) 
#>   ..$ AEO_0000129 : chr(0) 
#>   ..$ AEO_0000131 : chr(0) 
#>   ..$ CL_0000034  : chr(0) 
#>   ..$ AEO_0000132 : chr "AEO_0000214"
#>   ..$ AEO_0000135 : chr [1:2] "AEO_0001000" "AEO_0001001"
#>   ..$ AEO_0000136 : chr(0) 
#>   ..$ AEO_0001012 : chr [1:2] "AEO_0000136" "AEO_0001013"
#>   ..$ CL_0000127  : chr(0) 
#>   ..$ CL_0000129  : chr(0) 
#>   ..$ CL_0000128  : chr "CL_0000127"
#>   ..$ AEO_0000137 : chr [1:2] "AEO_0000201" "AEO_0000202"
#>   ..$ AEO_0000138 : chr(0) 
#>   ..$ CL_0000526  : chr "CL_0000101"
#>   ..$ CL_0000527  : chr(0) 
#>   ..$ AEO_0000139 : chr(0) 
#>   ..$ AEO_0001011 : chr "AEO_0000139"
#>   ..$ AEO_0000140 : chr "AEO_0000144"
#>   ..$ CL_0000737  : chr [1:3] "CL_0000746" "CL_0000444" "CL_0002372"
#>   ..$ AEO_0000141 : chr "AEO_0000143"
#>   ..$ CL_0000192  : chr(0) 
#>   ..$ AEO_0000142 : chr(0) 
#>   ..$ CL_0000746  : chr(0) 
#>   ..$ AEO_0000143 : chr(0) 
#>   ..$ AEO_0000144 : chr(0) 
#>   ..$ AEO_0000145 : chr [1:3] "AEO_0000146" "AEO_0000151" "AEO_0000152"
#>   ..$ AEO_0000146 : chr "AEO_0000148"
#>   ..$ AEO_0000147 : chr(0) 
#>   ..$ AEO_0000148 : chr [1:4] "AEO_0000147" "AEO_0000149" "AEO_0000150" "AEO_0000213"
#>   ..$ AEO_0001016 : chr [1:4] "AEO_0000148" "AEO_0000212" "AEO_0001017" "AEO_0001018"
#>   ..$ CL_0000335  : chr(0) 
#>   ..$ AEO_0000149 : chr(0) 
#>   ..$ CL_0000680  : chr(0) 
#>   ..$ AEO_0000150 : chr(0) 
#>   ..$ AEO_0000151 : chr(0) 
#>   ..$ AEO_0000152 : chr "AEO_0000212"
#>   ..$ CL_0000134  : chr "CL_0000335"
#>   ..$ AEO_0000153 : chr(0) 
#>   ..$ AEO_0000187 : chr [1:2] "AEO_0000153" "AEO_0000189"
#>   ..$ AEO_0000154 : chr(0) 
#>   .. [list output truncated]
#>  $ ancestors:List of 250
#>   ..$ AEO_0000013 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CARO_0000043: chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043"
#>   ..$ AEO_0000078 : chr [1:5] "CARO_0000000" "CARO_0000007" "CARO_0000005" "AEO_0000221" ...
#>   ..$ AEO_0000221 : chr [1:4] "CARO_0000000" "CARO_0000007" "CARO_0000005" "AEO_0000221"
#>   ..$ AEO_0000079 : chr [1:5] "CARO_0000000" "CARO_0000007" "CARO_0000005" "AEO_0000222" ...
#>   ..$ AEO_0000222 : chr [1:4] "CARO_0000000" "CARO_0000007" "CARO_0000005" "AEO_0000222"
#>   ..$ AEO_0000080 : chr [1:4] "CARO_0000000" "CARO_0000007" "CARO_0000010" "AEO_0000080"
#>   ..$ CARO_0000010: chr [1:3] "CARO_0000000" "CARO_0000007" "CARO_0000010"
#>   ..$ AEO_0000081 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000082 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ CARO_0000055: chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055"
#>   ..$ AEO_0000083 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ CL_0000138  : chr [1:11] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ CL_0000058  : chr [1:12] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000055" ...
#>   ..$ CL_0000062  : chr [1:8] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000055" ...
#>   ..$ AEO_0000168 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000054" ...
#>   ..$ AEO_0000085 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000086 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000087 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000090 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000327  : chr [1:9] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ CL_0000057  : chr [1:7] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000091 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000092 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0002320  : chr [1:6] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000093 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000054" ...
#>   ..$ CARO_0000041: chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000054" ...
#>   ..$ AEO_0000094 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000024" ...
#>   ..$ CARO_0000024: chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000024"
#>   ..$ AEO_0000095 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000096 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000097 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000098 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000099 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000040" ...
#>   ..$ CARO_0000040: chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000040"
#>   ..$ AEO_0000100 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000040" ...
#>   ..$ AEO_0000103 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032" ...
#>   ..$ CARO_0000032: chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032"
#>   ..$ AEO_0000106 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032" ...
#>   ..$ AEO_0000108 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032" ...
#>   ..$ AEO_0000109 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032" ...
#>   ..$ AEO_0000193 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032" ...
#>   ..$ AEO_0000111 : chr [1:8] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CARO_0000070: chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000114 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CARO_0000066: chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000068  : chr [1:8] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000063" ...
#>   ..$ AEO_0000115 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000075  : chr [1:8] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000063" ...
#>   ..$ AEO_0000117 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000118 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000119f: chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000122 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000123 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000125 : chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "AEO_0000125"
#>   ..$ CARO_0000003: chr [1:3] "CARO_0000000" "CARO_0000006" "CARO_0000003"
#>   ..$ AEO_0000126 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000012" ...
#>   ..$ CARO_0000012: chr [1:4] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000012"
#>   ..$ AEO_0000127 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "AEO_0000125" ...
#>   ..$ CL_0000445  : chr [1:5] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000129 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "AEO_0000125" ...
#>   ..$ AEO_0000131 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "AEO_0000125" ...
#>   ..$ CL_0000034  : chr [1:5] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000132 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "AEO_0000125" ...
#>   ..$ AEO_0000135 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000136 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0001012 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000127  : chr [1:12] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ CL_0000129  : chr [1:10] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ CL_0000128  : chr [1:11] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000137 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000138 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000526  : chr [1:8] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ CL_0000527  : chr [1:8] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000139 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0001011 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000140 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000737  : chr [1:9] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000141 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000192  : chr [1:9] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000142 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000746  : chr [1:10] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000144" ...
#>   ..$ AEO_0000143 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000144 : chr [1:7] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000145 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000146 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000147 : chr [1:10] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000148 : chr [1:9] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0001016 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "AEO_0000125" ...
#>   ..$ CL_0000335  : chr [1:6] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000063" ...
#>   ..$ AEO_0000149 : chr [1:10] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000680  : chr [1:7] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000055" ...
#>   ..$ AEO_0000150 : chr [1:10] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000151 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ AEO_0000152 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000043" ...
#>   ..$ CL_0000134  : chr [1:5] "CL_0000000" "CL_0000003" "CL_0000012" "CL_0000063" ...
#>   ..$ AEO_0000153 : chr [1:6] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000187 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000055" ...
#>   ..$ AEO_0000154 : chr [1:5] "CARO_0000000" "CARO_0000006" "CARO_0000003" "CARO_0000032" ...
#>   .. [list output truncated]
#>  $ obsolete : Named logi [1:250] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..- attr(*, "names")= chr [1:250] "AEO_0000013" "CARO_0000043" "AEO_0000078" "AEO_0000221" ...
#>  - attr(*, "class")= chr "ontology_index"