preparing for a small number of entry points to owlready2 mediated by basilisk, this setup function will ingest OWL, enumerate classes and their names, and produce the 'parents' list, which can then be used with ontology_index to produce a functional ontology representation

preparing for a small number of entry points to owlready2 mediated by basilisk, this setup function will ingest OWL, enumerate classes and their names, and produce the 'parents' list, which can then be used with ontology_index to produce a functional ontology representation

setup_entities2(owlfn)

Arguments

owlfn

character(1) path to OWL file

Examples

pa = get_ordo_owl_path()
orde = setup_entities2(pa)
orde
#> Ontology with 15239 terms
#> 
#> Properties:
#> 	id: character
#> 	name: character
#> 	parents: list
#> 	children: list
#> 	ancestors: list
#> 	obsolete: logical
#> Roots:
#> 	Orphanet_C010 - genetic material
#> 	Orphanet_C001 - clinical entity
#> 	Orphanet_C041 - inactive clinical entity
#> 	4.0 - CC BY 4.0