generate a TermSet with siblings of a given term, excluding that term by default

generate a TermSet with siblings of a given term, excluding that term by default

acquire the label of an ontology subject tag

acquire the labels of children of an ontology subject tag

siblings_TAG(Tagstring = "EFO:1001209", ontology, justSibs = TRUE)

label_TAG(Tagstring = "EFO:0000311", ontology)

children_TAG(Tagstring = "EFO:1001209", ontology)

Arguments

Tagstring
a character(1) that identifies a term

ontology
instance of ontology_index (S3) from ontologyIndex

justSibs
character(1)

Value

TermSet instance

character(1)

TermSet instance

Note

for label_TAG, Tagstring may be a vector

Examples

efoOnto = getEFOOnto()
siblings_TAG( "EFO:1001209", efoOnto )

#> TermSet for 4 terms
#> Churg-Strauss syndrome, microscopic polyangiitis, retinal vasculitis, Takayasu arteritis 

efoOnto = getEFOOnto()
label_TAG( "EFO:0000311", efoOnto )

#> EFO:0000311 
#>    "cancer" 

efoOnto = getEFOOnto()
children_TAG( ontology = efoOnto )

#> TermSet for 0 terms
#>