subset a SummarizedExperiment to which ontology tags have been bound using 'bind_formal_tags', obtaining the 'descendants' of the class of interest

subset a SummarizedExperiment to which ontology tags have been bound using 'bind_formal_tags', obtaining the 'descendants' of the class of interest

subset_descendants(
  se,
  onto,
  class_name,
  class_tag,
  formal_cd_name = "label.ont"
)

Arguments

se
SummarizedExperiment instance

onto
representation of an ontology using representation from ontologyIndex package

class_name
character(1) if 'class_tag' is missing, this will be grepped in onto[["name"]] to find class and its descendants

class_tag
character(1) used if given to identify "ontological descendants" of this term in se

formal_cd_name
character(1) tells name used for ontology tag column in `colData(se)`

Value

instance of SummarizedExperiment