use Cell Ontology and Protein Ontology to identify cell-type defining conditions in which a given gene is named

use Cell Ontology and Protein Ontology to identify cell-type defining conditions in which a given gene is named

sym2CellOnto(sym, cl, pr)

Arguments

sym
gene symbol, must be used in protein ontology as a PRO:DNx exact match token

cl
result of getCellOnto()

pr
result of getPROnto()

Value

DataFrame if any hits are found. A field 'cond' abbreviates the identified conditions: (has/lacks)PMP (plasma membrane part) (hi/lo)PMAmt (plasma membrane amount), (has/lacks)Part.

Note

Currently just checks for *plasma_membrane_part, *plasma_membrane_amount, and *Part conditions.

Examples

if (!exists("cl")) cl = getCellOnto()
if (!exists("pr")) pr = getPROnto()
sym2CellOnto("ITGAM", cl, pr)

#> DataFrame with 39 rows and 4 columns
#>             sym        cond          cl                   type
#>     <character> <character> <character>            <character>
#> 1         ITGAM    lacksPMP  CL:0000037 hematopoietic stem c..
#> 2         ITGAM    lacksPMP  CL:0000547        proerythroblast
#> 3         ITGAM    lacksPMP  CL:0000553 megakaryocyte progen..
#> 4         ITGAM    lacksPMP  CL:0000558           reticulocyte
#> 5         ITGAM    lacksPMP  CL:0000611 eosinophil progenito..
#> ...         ...         ...         ...                    ...
#> 35        ITGAM     loPMAmt  CL:0002345 CD27-low, CD11b-low ..
#> 36        ITGAM     loPMAmt  CL:0002349 CD27-high, CD11b-low..
#> 37        ITGAM     loPMAmt  CL:0002506 liver CD103-positive..
#> 38        ITGAM     loPMAmt  CL:0002509 CD103-positive, lang..
#> 39        ITGAM     loPMAmt  CL:0002511 CD11b-low, CD103-neg..

sym2CellOnto("FOXP3", cl, pr)

#> DataFrame with 4 rows and 4 columns
#>           sym        cond          cl                   type
#>   <character> <character> <character>            <character>
#> 1       FOXP3     hasPart  CL:0000902 induced T-regulatory..
#> 2       FOXP3     hasPart  CL:0000903 natural T-regulatory..
#> 3       FOXP3     hasPart  CL:0000919 CD8-positive, CD25-p..
#> 4       FOXP3     hasPart  CL:0000920 CD8-positive, CD28-n..