produce a data.frame of features relevant to a Cell Ontology class

produce a data.frame of features relevant to a Cell Ontology class

Usage

CLfeats(ont, tag = "CL:0001054", pr, go)

Arguments

ont

instance of ontologyIndex ontology

tag

character(1) a CL: class tag

pr

instance of ontologyIndex PRO protein ontology

go

instance of ontologyIndex GO gene ontology

Value

a data.frame instance

Note

This function will look in the intersection_of and has_part, lacks_part components of the CL entry to find properties asserted of or inherited by the cell type identified in 'tag'. As of 1.19, this function does not look in global environment for ontologies. We use 2021 versions in the examples because some changes in ontologies omit important relationships; revisions to package code after 1.19.4 will attempt to address these.

Examples

cl = getOnto("cellOnto", year_added="2021")
#> loading from cache
pr = getOnto("Pronto", "2021")  # legacy tag, for 2022 would be PROnto
#> loading from cache
go = getOnto("goOnto", "2021")
#> loading from cache
CLfeats(cl, tag="CL:0001054", pr=pr, go=go)
#> no recognized predicate references for CL:0000576
#> no recognized predicate references for CL:0000766
#> no recognized predicate references for CL:0000738
#> no recognized predicate references for CL:0000988
#> no recognized predicate references for CL:0000219
#> no recognized predicate references for CL:0000003
#>          tag        prtag     cond
#> 1 CL:0001054 PR:000001002 lacksPMP
#> 2 CL:0001054 PR:000001020 lacksPMP
#> 3 CL:0001054 PR:000001289 lacksPMP
#> 4 CL:0000000   GO:0005634  hasPart
#> 5 CL:0000000   GO:0005634  hasPart
#>                                             entity SYMBOL
#> 1                                    CD19 molecule   CD19
#> 2                                      CD3 epsilon   CD3E
#> 3 membrane-spanning 4-domains subfamily A member 1  MS4A1
#> 4                                          nucleus   <NA>
#> 5                                          nucleus   <NA>
#>                     name
#> 1 CD14-positive monocyte
#> 2 CD14-positive monocyte
#> 3 CD14-positive monocyte
#> 4                   cell
#> 5                   cell