utilities for approximate matching of cell type terms to GO categories and annotations

utilities for approximate matching of cell type terms to GO categories and annotations

Usage

cellTypeToGO(celltypeString, gotab, ...)

cellTypeToGenes(
  celltypeString,
  gotab,
  orgDb,
  cols = c("ENSEMBL", "SYMBOL"),
  ...
)

Arguments

celltypeString

character atom to be used to search GO terms using

gotab

a data.frame with columns GO (goids) and TERM (term strings) agrep

...

additional arguments to agrep

orgDb

instances of orgDb

cols

columns to be retrieved in select operation

Value

data.frame

data.frame

Note

Very primitive, uses agrep to try to find relevant terms.

Examples

library(org.Hs.eg.db)
#> Loading required package: AnnotationDbi
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, table, tapply,
#>     union, unique, unsplit, which.max, which.min
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> Loading required package: IRanges
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> 
data(allGOterms)
head(cellTypeToGO("serotonergic neuron", allGOterms))
#>             GOID                                        TERM
#> 18623 GO:0036515           serotonergic neuron axon guidance
#> 18625 GO:0036517 chemoattraction of serotonergic neuron axon
#> 18627 GO:0036519  chemorepulsion of serotonergic neuron axon
head(cellTypeToGenes("serotonergic neuron", allGOterms, org.Hs.eg.db))
#> 'select()' returned 1:many mapping between keys and columns
#>           GO EVIDENCE ONTOLOGY         ENSEMBL  SYMBOL
#> 1 GO:0036515      ISS       BP ENSG00000008300  CELSR3
#> 2 GO:0036515      ISS       BP ENSG00000104290    FZD3
#> 3 GO:0036515      IEA       BP ENSG00000243232 PCDHAC2
#> 4 GO:0036515      IEA       BP ENSG00000291668 PCDHAC2
#> 5 GO:0036515      ISS       BP ENSG00000162738  VANGL2
#> 6 GO:0036517      ISS       BP ENSG00000114251   WNT5A