load ontologies that may include non-ascii strings and therefore cannot be in data folder

load ontologies that may include non-ascii strings and therefore cannot be in data folder

getCellOnto(
  useNew = TRUE,
  newest = FALSE,
  cache = BiocFileCache::BiocFileCache(),
  use0718 = FALSE
)

getCellLineOnto()

getEFOOnto()

getChebiLite()

getCellosaurusOnto()

getUBERON_NE()

getChebiOnto()

getOncotreeOnto()

getDiseaseOnto()

getGeneOnto()

getHCAOnto()

getPROnto()

getPATOnto()

getMondoOnto()

getSIOOnto()

Arguments

useNew

logical(1) only for getCellOnto if TRUE return ontology_index instance of cell ontology 2.1 of May 21 2020, defaults to TRUE

newest

logical(1) if TRUE will use BiocFileCache to retrieve/use latest cl-simple.obo; overrides

cache

instance of BiocFileCache

use0718

logical(1) only for getCellOnto if TRUE cell ontology of July 2018

Value

instance of ontology_index (S3) from ontologyIndex

instance of ontology_index (S3) from ontologyIndex

Note

You may want to try `bfcupdate` on the BiocFileCache element. useNew

Provenance information is kept in the form of excerpts of top records in `dir(system.file("obo", package="ontoProc"), full=TRUE)`. This function is deprecated in Bioc 3.16. Use getOnto("cellOnto") to get the latest version provided by AnnotationHub.

getChebiOnto loads ontoRda/chebi_full.rda

getOncotreeOnto loads ontoRda/oncotree.rda

getDiseaseOnto loads ontoRda/diseaseOnto.rda

getHCAOnto loads ontoRda/hcaOnto.rda produced from hcao.owl at https://github.com/HumanCellAtlas/ontology/releases/tag/1.0.6 2/11/2019, python pronto was used to convert OWL to OBO.

getPROnto loads ontoRda/PRonto.rda, produced from http://purl.obolibrary.org/obo/pr.obo 'reasoned' ontology from OBO foundry, 02-08-2019. In contrast to other ontologies, this is imported via get_OBO with `extract_tags='minimal'`.

getPATOnto loads ontoRda/patoOnto.rda, produced from https://raw.githubusercontent.com/pato-ontology/pato/master/pato.obo from OBO foundry, 02-08-2019.

Examples

co = getCellOnto(useNew=TRUE)
#> Warning: getCellOnto is deprecated: getOnto('cellOnto') should be used for versioned access
co
#> Ontology with 9941 terms
#> 
#> format-version: 1.2
#> data-version: releases/2021-03-05
#> ontology: cl
#> 
#> Properties:
#> 	id: character
#> 	name: character
#> 	parents: list
#> 	children: list
#> 	ancestors: list
#> 	obsolete: logical
#> 	adjacent_to: list
#> 	alt_id: list
#> 	anterior_to: list
#> 	anteriorly_connected_to: list
#> 	attaches_to: list
#> 	attaches_to_part_of: list
#> 	bearer_of: list
#> 	bounding_layer_of: list
#> 	branching_part_of: list
#> 	capable_of: list
#> 	capable_of_part_of: list
#> 	channel_for: list
#> 	channels_from: list
#> 	channels_into: list
#> 	comment: character
#> 	composed_primarily_of: list
#> 	conduit_for: list
#> 	connected_to: list
#> 	connects: list
#> 	consider: list
#> 	contains: list
#> 	continuous_with: list
#> 	contributes_to_morphology_of: list
#> 	created_by: character
#> 	creation_date: character
#> 	data-version: list
#> 	dc-contributor: list
#> 	dc-creator: list
#> 	decreased_in_magnitude_relative_to: list
#> 	deep_to: list
#> 	def: character
#> 	derives_from: list
#> 	developmentally_induced_by: list
#> 	developmentally_replaces: list
#> 	develops_from: list
#> 	develops_from_part_of: list
#> 	develops_in: list
#> 	develops_into: list
#> 	directly_develops_from: list
#> 	disjoint_from: list
#> 	distal_to: list
#> 	distally_connected_to: list
#> 	distalmost_part_of: list
#> 	domain: list
#> 	dorsal_to: list
#> 	drains: list
#> 	ends: list
#> 	ends_with: list
#> 	equivalent_to_chain: list
#> 	existence_ends_during: list
#> 	existence_ends_during_or_before: list
#> 	existence_ends_with: list
#> 	existence_starts_and_ends_during: list
#> 	existence_starts_during: list
#> 	existence_starts_during_or_after: list
#> 	existence_starts_with: list
#> 	expand_assertion_to: list
#> 	expand_expression_to: list
#> 	extends_fibers_into: list
#> 	filtered_through: list
#> 	format-version: list
#> 	has_boundary: list
#> 	has_completed: list
#> 	has_component: list
#> 	has_cross_section: list
#> 	has_developmental_contribution_from: list
#> 	has_gene_template: list
#> 	has_high_plasma_membrane_amount: list
#> 	has_low_plasma_membrane_amount: list
#> 	has_member: list
#> 	has_muscle_antagonist: list
#> 	has_muscle_insertion: list
#> 	has_muscle_origin: list
#> 	has_not_completed: list
#> 	has_part: list
#> 	has_potential_to_develop_into: list
#> 	has_potential_to_developmentally_contribute_to: list
#> 	has_quality: list
#> 	has_relative_magnitude: list
#> 	has_role: list
#> 	has_skeleton: list
#> 	holds_over_chain: list
#> 	immediate_transformation_of: list
#> 	immediately_deep_to: list
#> 	immediately_preceded_by: list
#> 	immediately_superficial_to: list
#> 	in_anterior_side_of: list
#> 	in_deep_part_of: list
#> 	in_distal_side_of: list
#> 	in_dorsal_side_of: list
#> 	in_lateral_side_of: list
#> 	in_left_side_of: list
#> 	in_posterior_side_of: list
#> 	in_proximal_side_of: list
#> 	in_right_side_of: list
#> 	in_taxon: list
#> 	increased_in_magnitude_relative_to: list
#> 	indirectly_supplies: list
#> 	innervated_by: list
#> 	innervates: list
#> 	intersection_of: list
#> 	intersects_midsagittal_plane_of: list
#> 	inverse_of: list
#> 	is_a: list
#> 	is_class_level: list
#> 	is_cyclic: list
#> 	is_functional: list
#> 	is_metadata_tag: list
#> 	is_symmetric: list
#> 	is_transitive: list
#> 	lacks_part: list
#> 	lacks_plasma_membrane_part: list
#> 	located_in: list
#> 	location_of: list
#> 	lumen_of: list
#> 	luminal_space_of: list
#> 	namespace: list
#> 	negatively_regulates: list
#> 	never_in_taxon: list
#> 	occurs_in: list
#> 	only_in_taxon: list
#> 	ontology: list
#> 	output_of: list
#> 	part_of: list
#> 	participates_in: list
#> 	positively_regulates: list
#> 	posterior_to: list
#> 	posteriorly_connected_to: list
#> 	preaxialmost_part_of: list
#> 	preceded_by: list
#> 	precedes: list
#> 	produced_by: list
#> 	produces: list
#> 	property_value: list
#> 	protects: list
#> 	proximal_to: list
#> 	proximally_connected_to: list
#> 	proximalmost_part_of: list
#> 	range: list
#> 	regulates: list
#> 	remark: list
#> 	replaced_by: list
#> 	RO:0002100: list
#> 	RO:0002102: list
#> 	RO:0002103: list
#> 	RO:0002104: list
#> 	RO:0002120: list
#> 	RO:0002130: list
#> 	RO:0002292: list
#> 	RO:0013001: list
#> 	seeAlso: list
#> 	sexually_homologous_to: list
#> 	skeleton_of: list
#> 	starts: list
#> 	starts_with: list
#> 	subdivision_of: list
#> 	subset: list
#> 	subsetdef: list
#> 	superficial_to: list
#> 	supplies: list
#> 	surrounded_by: list
#> 	surrounds: list
#> 	synonym: list
#> 	synonymtypedef: list
#> 	transformation_of: list
#> 	transitive_over: list
#> 	tributary_of: list
#> 	trunk_part_of: list
#> 	union_of: list
#> 	ventral_to: list
#> 	xref: list
#> Roots:
#> 	BFO:0000003 - occurrent
#> 	functionally_related_to - functionally related to
#> 	CHEBI:36342 - subatomic particle
#> 	RO:0002410 - causally related to
#> 	BFO:0000002 - continuant
#> 	UBERON:0001062 - anatomical entity
#> 	RO:0002323 - mereotopologically related to
#> 	GO:0005575 - cellular_component
#> 	UBERON:0000000 - processual entity
#> 	RO:0000057 - has participant
#>  ... 82 more
clo = getCellLineOnto()
#> Warning: cellLineOnto is deprecated: getOnto('cellLineOnto') should be used for versioned access.
#> snapshotDate(): 2022-10-19
#> loading from cache
length(clo$id)
#> [1] 39192
che = getChebiLite()
#> snapshotDate(): 2022-10-19
#> loading from cache
length(che$id)
#> [1] 137347
efo = getEFOOnto()
#> Warning: efoOnto is deprecated: getOnto('efoOnto') should be used for versioned access.
#> snapshotDate(): 2022-10-19
#> loading from cache
length(efo$id)
#> [1] 28424