use Cell Ontology and Protein Ontology to identify cell-type defining conditions in which a given gene is named

use Cell Ontology and Protein Ontology to identify cell-type defining conditions in which a given gene is named

Usage

sym2CellOnto(sym, cl, pr)

Arguments

sym

gene symbol, must be used in protein ontology as a PRO:DNx exact match token

cl

result of getOnto("cellOnto")

pr

result of getOnto("PROnto")

Value

DataFrame if any hits are found. A field 'cond' abbreviates the identified conditions: (has/lacks)PMP (plasma membrane part) (hi/lo)PMAmt (plasma membrane amount), (has/lacks)Part.

Note

Currently just checks for *plasma_membrane_part, *plasma_membrane_amount, and *Part conditions.

Examples

if (!exists("cl")) cl = getOnto("cellOnto")
#> loading from cache
if (!exists("pr")) pr = getOnto("PROnto")
#> loading from cache
sym2CellOnto("ITGAM", cl, pr)
#> DataFrame with 22 rows and 4 columns
#>             sym        cond          cl                   type
#>     <character> <character> <character>            <character>
#> 1         ITGAM    lacksPMP  CL:0000037 hematopoietic stem c..
#> 2         ITGAM    lacksPMP  CL:0000547        proerythroblast
#> 3         ITGAM    lacksPMP  CL:0000553 megakaryocyte progen..
#> 4         ITGAM    lacksPMP  CL:0000558           reticulocyte
#> 5         ITGAM    lacksPMP  CL:0000611 eosinophil progenito..
#> ...         ...         ...         ...                    ...
#> 18        ITGAM    lacksPMP  CL:0001060 hematopoietic oligop..
#> 19        ITGAM    lacksPMP  CL:0001066 erythroid progenitor..
#> 20        ITGAM    lacksPMP  CL:0002010 pre-conventional den..
#> 21        ITGAM    lacksPMP  CL:0002089 group 2 innate lymph..
#> 22        ITGAM    lacksPMP  CL:0002679 natural helper lymph..
sym2CellOnto("FOXP3", cl, pr)
#> DataFrame with 4 rows and 4 columns
#>           sym        cond          cl                   type
#>   <character> <character> <character>            <character>
#> 1       FOXP3     hasPart  CL:0000902 induced T-regulatory..
#> 2       FOXP3     hasPart  CL:0000903 natural T-regulatory..
#> 3       FOXP3     hasPart  CL:0000919 CD8-positive, CD25-p..
#> 4       FOXP3     hasPart  CL:0000920 CD8-positive, CD28-n..